requirements.txt
Pillow==1.7.5
python-memcached==1.48
pymongo
git+https://github.com/django-nonrel/mongodb-engine.git#egg=django_mongodb_engine
hg+http://bitbucket.org/wkornewald/django-nonrel#egg=Django
hg+http://bitbucket.org/wkornewald/djangotoolbox#egg=djangotoolbox
hg+http://bitbucket.org/fragro/django-customtags-lib#egg=customtags
hg+http://bitbucket.org/fragro/django-tagging-multidb#egg=tagging
hg+http://bitbucket.org/offline/django-annoying#egg=annoying
hg+http://bitbucket.org/fragro/python-markdown#egg=markdown
hg+http://bitbucket.org/twanschik/django-autoload#egg=autoload
hg+http://bitbucket.org/wkornewald/django-dbindexer#egg=dbindexer
hg+http://bitbucket.org/wkornewald/django-filetransfers#egg=filetransfers
oauth2
httplib2
python-openid
simplejson
django-piston
pyorbited
hg+http://bitbucket.org/fragro/django-tracking#egg=tracking
hg+http://bitbucket.org/nator/django-markitup#egg=markitup
hg+http://bitbucket.org/twanschik/nonrel-search#egg=search
django-celery
setproctitle
pytz
git+https://github.com/bradbeattie/python-vote-core.git#egg=pyvotecore
git+https://github.com/fragro/django-notification.git#egg=notification
johnny-cache
git+https://github.com/sorl/sorl-thumbnail.git#egg=sorl
